+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `PreciseDateTime` class to indicate DateTime with milliseconds'
 - Changed `AccountRequest.requestTime` and `AccountRequest.responseTime` type to `PreciseDateTime`
 - Update `Branding` and `AccountBilling` resource
+- Add `AccountMe.changePasswordRequest` endpoint
 
 ## [2.8.1] - 2025-06-23
 - Rename `AccountBilling.fileUploadLimit` to `AccountBilling.fileSizeLimit`